+++ halotools/empirical_models/__init__.py
@@ -14,8 +14,8 @@ from .abunmatch import *
 from .model_helpers import *
 from .assembias_models import *
 from .hod_designer import *
-from .conc_mass_models import *
-from .monte_carlo_phase_space import *
+# from .conc_mass_models import *
+# from .monte_carlo_phase_space import *
 from .phase_space_models import *
 # from .profile_helpers import *
 from .profile_models import *
