+++ halotools/empirical_models/phase_space_models.py
@@ -15,7 +15,7 @@ from astropy.table import Table
 
 from .profile_models import *
 from .velocity_models import *
-from .monte_carlo_phase_space import *
+from .phase_space_marf.monte_carlo_phase_space import *
 from . import model_defaults
 
 from ..sim_manager import sim_defaults
