+++ halotools/empirical_models/profile_models.py
@@ -17,7 +17,7 @@ from astropy.constants import G
 newtonG = G.to(u.km*u.km*u.Mpc/(u.Msun*u.s*u.s))
 
 from . import model_defaults
-from .conc_mass_models import ConcMass
+from .phase_space_marf.conc_mass_models import ConcMass
 from .phase_space_marf.profile_helpers import *
 
 from ..utils.array_utils import convert_to_ndarray
